Switch to TS project references for simpler project builds
Test project must be built separately and is not built by default. This is due to it depending on the server part of the project which causes things to double build. Also, tests don't need to build for the normal checkout + build process so this speeds Docker builds up a bit.